myp/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-ipware==3.0.2
django-qr-code==2.0.1
pytz==2020.4
segno==1.3.1
sqlparse==0.4.1
